backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
sentence-transformers>=2.7.0
huggingface_hub>=0.20.0
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
pytesseract==0.3.10
Pillow>=10.0.0
pdf2image==1.16.3
scikit-learn==1.3.2
spacy==3.7.2
openai-whisper==20231117
coqui-tts==0.27.3
numpy>=1.26.0,<2.0
grpcio>=1.48.2
contourpy>=1.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
soundfile>=0.12.1
scipy>=1.11.0

